command: commands: Add macros to declare commands with subcmds
Most cmd/xxx.c source files expose several commands through a single
entry point. Some of them are doing the sub-command parsing manually in
their do_<cmd>() function, others are declaring a table of sub-commands
and then use find_cmd_tbl() to delegate the request to the sub command
handler.

In both case, the amount of code to do that is not negligible and
repetitive, not to mention that almost no commands are implementing
a auto-completion hook, which means most u-boot lack auto-completion.

Provide several macros to easily define sub-commands and commands
exposing such sub-commands.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>